Transaction id type changed. Article crosses added foregion keys to article table.

+++ src/Services/Main/Main.Application/Handlers/ArticlePairs/DeletePair/DeletePairHandler.cs
@@ -1,12 +1,14 @@
 using Application.Common.Interfaces;
 using Core.Attributes;
 using Core.Interfaces.Services;
+using Exceptions.Exceptions.ArticlePair;
 using Main.Core.Interfaces.DbRepositories;
 using MediatR;
 
 namespace Main.Application.Handlers.ArticlePairs.DeletePair;
 
 [Transactional]
+[ExceptionType<ArticlePairNotFoundException>]
 public record DeletePairCommand(int ArticleId) : ICommand;
 
 public class DeletePairHandler(IArticlePairsRepository pairsRepository, IUnitOfWork unitOfWork)
@@ -14,8 +16,12 @@ public class DeletePairHandler(IArticlePairsRepository pairsRepository, IUnitOfW
 {
     public async Task<Unit> Handle(DeletePairCommand request, CancellationToken cancellationToken)
     {
-        var pairs = await pairsRepository
-            .GetRelatedPairsAsync(request.ArticleId, true, cancellationToken);
+        var pairs = (await pairsRepository
+            .GetRelatedPairsAsync(request.ArticleId, true, cancellationToken))
+            .ToList();
+
+        if (pairs.Count == 0) throw new ArticlePairNotFoundException(request.ArticleId);
+        
         unitOfWork.RemoveRange(pairs);
         await unitOfWork.SaveChangesAsync(cancellationToken);
         return Unit.Value;
